создал все сущности для обмена в контроллерах

+++ src/main/java/ru/skypro/homework/dto/Ad.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Ad {
 
-    private Integer author;     // id автора объявления
+    private Long author;        // id автора объявления
     private String image;       // ссылка на картинку объявления
     private Integer pk;         // id объявления
     private Integer price;      // цена объявления
